Added new command and extended two others
Extended output of "stafflist" to include two more op/admin roles and added permissions to define the two roles.

Added new command "amenu" to show how to make an admin only menu. Included is the permission needed to use the command to view the menu as well as samples for admin only commands.

Extended the default inventory "menu" to include "friends list" and "friends settings" as options to better show how to pass arguments with commands via the gui and give new server owners a good springboard for setting up a menu.